dom/workers/test/test_worker_interfaces_secureContext.html

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->
<!DOCTYPE HTML>
<html>
<head>
  <title>Validate Interfaces Exposed to Workers</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="worker_driver.js"></script>
</head>
<body>
<script>
  ok(self.isSecureContext, "This test should be running in a secure context");
</script>
<script class="testbody" type="text/javascript">
workerTestExec("test_worker_interfaces.js");
</script>
</body>
</html>
